requirements.txt
requests
pyyaml
python-dateutil
openai
